--- specifications/src/main/java/org/dcsa/conformance/specifications/standards/ct/v300/messages/GetEventsResponse.java
@@ -1,29 +1,0 @@
-package org.dcsa.conformance.specifications.standards.ct.v300.messages;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import java.util.List;
-import lombok.Data;
-import org.dcsa.conformance.specifications.standards.ct.v300.model.ContainerTrackingEvent;
-
-@Data
-@Schema(
-    description =
-"""
-API response to a `GET Events` request
-""")
-public class GetEventsResponse {
-
-  @Schema(
-      description =
-"""
-The list of events matching the filter specified by the request query parameters
-""")
-  private List<ContainerTrackingEvent> events;
-
-  @Schema(
-      description =
-"""
-Optional list of feedback elements indicating how the request was processed
-""")
-  private List<FeedbackElement> feedbackElements;
-}
