--- specifications/src/main/java/org/dcsa/conformance/specifications/standards/ct/v300/model/EquipmentEventInfo.java
@@ -1,12 +1,0 @@
-package org.dcsa.conformance.specifications.standards.ct.v300.model;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Data;
-import org.dcsa.conformance.specifications.standards.ct.v300.types.EquipmentEventTypeCode;
-
-@Data
-@Schema(description = "Equipment event information")
-public class EquipmentEventInfo {
-
-  @Schema() private EquipmentEventTypeCode equipmentEventTypeCode;
-}
