--- specifications/src/main/java/org/dcsa/conformance/specifications/standards/ct/v300/types/IotEventTypeCode.java
@@ -1,18 +1,0 @@
-package org.dcsa.conformance.specifications.standards.ct.v300.types;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.dcsa.conformance.specifications.generator.EnumBase;
-
-@Getter
-@Schema(
-    type = "string",
-    example = "DETC",
-    description = "Code used to denote the type of an IoT event")
-@AllArgsConstructor
-public enum IotEventTypeCode implements EnumBase {
-  DETC("Detected");
-
-  private final String valueDescription;
-}
