+++ specifications/src/main/java/org/dcsa/conformance/specifications/standards/tnt/v300/messages/GetEventsResponse.java
@@ -1,0 +1,29 @@
+package org.dcsa.conformance.specifications.standards.tnt.v300.messages;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import java.util.List;
+import lombok.Data;
+import org.dcsa.conformance.specifications.standards.tnt.v300.model.TNTEvent;
+
+@Data
+@Schema(
+    description =
+"""
+API response to a `GET Events` request
+""")
+public class GetEventsResponse {
+
+  @Schema(
+      description =
+"""
+The list of events matching the filter specified by the request query parameters
+""")
+  private List<TNTEvent> events;
+
+  @Schema(
+      description =
+"""
+Optional list of feedback elements indicating how the request was processed
+""")
+  private List<FeedbackElement> feedbackElements;
+}
