+++ specifications/src/main/java/org/dcsa/conformance/specifications/standards/tnt/v300/model/ShipmentEventInfo.java
@@ -1,0 +1,12 @@
+package org.dcsa.conformance.specifications.standards.tnt.v300.model;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Data;
+import org.dcsa.conformance.specifications.standards.tnt.v300.types.ShipmentEventTypeCode;
+
+@Data
+@Schema(description = "Shipment event information")
+public class ShipmentEventInfo {
+
+  @Schema() private ShipmentEventTypeCode shipmentEventTypeCode;
+}
