+++ specifications/src/main/java/org/dcsa/conformance/specifications/standards/tnt/v300/types/TransportEventTypeCode.java
@@ -1,0 +1,19 @@
+package org.dcsa.conformance.specifications.standards.tnt.v300.types;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.dcsa.conformance.specifications.generator.EnumBase;
+
+@Getter
+@Schema(
+    type = "string",
+    example = "ARRI",
+    description = "Code used to denote the type of a transport event")
+@AllArgsConstructor
+public enum TransportEventTypeCode implements EnumBase {
+  ARRI("Arrived"),
+  DEPA("Departed");
+
+  private final String valueDescription;
+}
