--- specifications/src/main/resources/conformance/specifications/ct/v300/changelog.md
@@ -1,5 +1,0 @@
-# Container Tracking v3.0.0 Changelog
-
-## Snapshot v3.0.0-20250829-design
-
-Initial Container Tracking v3.0.0 design stage snapshot. (SD-2433)
