+++ specifications/src/main/resources/conformance/specifications/tnt/v300/changelog.md
@@ -1,0 +1,5 @@
+# Track and Trace v3.0.0 Changelog
+
+## Snapshot v3.0.0-20250912-design
+
+Initial Track and Trace v3.0.0 design stage snapshot. (SD-2436)
