+++ specifications/src/main/resources/conformance/specifications/tnt/v300/openapi-get-events-description.md
